// src/app/api/kosarica/model/kosaricaDTO.ts
/**
 * Kosarica
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ElementDTO } from './elementDTO';

export interface KosaricaDTO {
  id_uporabnik?: number;
  kosarica?: Array<ElementDTO>;
}